refactor: change test dependencies import
Remove imports statements from core that imported
dependencies from the cli.
