+++ core/plugins/platforms/github/mod.test.ts
@@ -1,5 +1,5 @@
+import { assertEquals, deepMerge } from "../../../tests/mod.ts";
 import { context } from "../../../../cli/src/plugin/mod.ts";
-import { assertEquals, deepMerge } from "../../../../cli/deps.ts";
 import { RenderedTemplate } from "../../../../cli/src/platform/plugin.ts";
 
 import { github } from "./mod.ts";
